test refactor: unified dirty_chai to a single file
