--- tests/unit/dirty-chai.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const chai = require('chai');
-const dirtyChai = require('dirty-chai');
-
-chai.use(dirtyChai);
-
-module.exports = chai;
