+++ tests/unit/middleware/log_request_test.js
@@ -2,7 +2,7 @@
 
 const mockery = require('mockery');
 const sinon = require('sinon');
-const expect = require('../dirty-chai').expect;
+const expect = require('../../dirty_chai').expect;
 
 const getLogRequestInstance = loggerStub => {
   mockery.registerMock('../util/logger', loggerStub);
